backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pymongo>=4.6.0
pydantic-settings>=2.2.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
aiofiles>=23.0.0
python-multipart>=0.0.9

# PostgreSQL + SQLAlchemy
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# APScheduler
apscheduler>=3.10.0

# Socket.io
python-socketio>=5.11.0
python-engineio>=4.9.0

# Image processing
Pillow>=10.0.0

# Email (OTP sending)
aiosmtplib>=3.0.0
